Add backend surveillance scoring and frontend event-group UX improvements.
This batches the current API/model/service refactors, new scoring and dedup utilities, UI route and labeling updates, script/test updates, and associated config/build file changes so the repository state is synchronized.

Made-with: Cursor

+++ app/db/models.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 from sqlalchemy import (
     JSON,
     Boolean,
@@ -77,15 +78,15 @@ class MarketSnapshot(Base):
     market_pk: Mapped[int] = mapped_column(ForeignKey("markets.id"), index=True)
     ts: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now(), index=True)
 
-    last_price_dollars: Mapped[float | None] = mapped_column(Numeric(12, 4), nullable=True)
-    yes_bid_dollars: Mapped[float | None] = mapped_column(Numeric(12, 4), nullable=True)
-    yes_ask_dollars: Mapped[float | None] = mapped_column(Numeric(12, 4), nullable=True)
-    no_bid_dollars: Mapped[float | None] = mapped_column(Numeric(12, 4), nullable=True)
-    no_ask_dollars: Mapped[float | None] = mapped_column(Numeric(12, 4), nullable=True)
-    volume_fp: Mapped[float | None] = mapped_column(Numeric(18, 2), nullable=True)
-    volume_24h_fp: Mapped[float | None] = mapped_column(Numeric(18, 2), nullable=True)
-    open_interest_fp: Mapped[float | None] = mapped_column(Numeric(18, 2), nullable=True)
-    liquidity_dollars: Mapped[float | None] = mapped_column(Numeric(18, 4), nullable=True)
+    last_price_dollars: Mapped[Decimal | None] = mapped_column(Numeric(12, 4), nullable=True)
+    yes_bid_dollars: Mapped[Decimal | None] = mapped_column(Numeric(12, 4), nullable=True)
+    yes_ask_dollars: Mapped[Decimal | None] = mapped_column(Numeric(12, 4), nullable=True)
+    no_bid_dollars: Mapped[Decimal | None] = mapped_column(Numeric(12, 4), nullable=True)
+    no_ask_dollars: Mapped[Decimal | None] = mapped_column(Numeric(12, 4), nullable=True)
+    volume_fp: Mapped[Decimal | None] = mapped_column(Numeric(18, 2), nullable=True)
+    volume_24h_fp: Mapped[Decimal | None] = mapped_column(Numeric(18, 2), nullable=True)
+    open_interest_fp: Mapped[Decimal | None] = mapped_column(Numeric(18, 2), nullable=True)
+    liquidity_dollars: Mapped[Decimal | None] = mapped_column(Numeric(18, 4), nullable=True)
 
     market: Mapped["Market"] = relationship(back_populates="snapshots")
 
@@ -98,9 +99,9 @@ class Trade(Base):
     trade_id: Mapped[str] = mapped_column(String(128), unique=True, index=True)
     ts: Mapped[datetime] = mapped_column(DateTime(timezone=True), index=True)
 
-    yes_price_dollars: Mapped[float | None] = mapped_column(Numeric(12, 4), nullable=True)
-    no_price_dollars: Mapped[float | None] = mapped_column(Numeric(12, 4), nullable=True)
-    count_fp: Mapped[float | None] = mapped_column(Numeric(18, 2), nullable=True)
+    yes_price_dollars: Mapped[Decimal | None] = mapped_column(Numeric(12, 4), nullable=True)
+    no_price_dollars: Mapped[Decimal | None] = mapped_column(Numeric(12, 4), nullable=True)
+    count_fp: Mapped[Decimal | None] = mapped_column(Numeric(18, 2), nullable=True)
     taker_side: Mapped[str | None] = mapped_column(String(3), nullable=True)
 
     received_at: Mapped[datetime] = mapped_column(
@@ -131,12 +132,12 @@ class BookEvent(Base):
     ts: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), nullable=True)
 
     side: Mapped[str] = mapped_column(String(3))
-    price_dollars: Mapped[float] = mapped_column(Numeric(12, 4))
+    price_dollars: Mapped[Decimal] = mapped_column(Numeric(12, 4))
 
     # Snapshot rows carry the absolute level size; delta rows carry the signed
     # delta (positive adds contracts, negative removes, zero removes the level).
-    size_fp: Mapped[float | None] = mapped_column(Numeric(18, 2), nullable=True)
-    delta_fp: Mapped[float | None] = mapped_column(Numeric(18, 2), nullable=True)
+    size_fp: Mapped[Decimal | None] = mapped_column(Numeric(18, 2), nullable=True)
+    delta_fp: Mapped[Decimal | None] = mapped_column(Numeric(18, 2), nullable=True)
 
     is_snapshot: Mapped[bool] = mapped_column(Boolean)
 
@@ -169,7 +170,7 @@ class Anomaly(Base):
     market_pk: Mapped[int] = mapped_column(ForeignKey("markets.id"), index=True)
     latest_snapshot_id: Mapped[int | None] = mapped_column(nullable=True, index=True)
 
-    score: Mapped[float] = mapped_column(Numeric(10, 2))
+    score: Mapped[Decimal] = mapped_column(Numeric(10, 2))
     severity: Mapped[str] = mapped_column(String(20), index=True)
     reasons: Mapped[list] = mapped_column(JSON)
     signals: Mapped[dict] = mapped_column(JSON)
